requirements.txt
canvasapi>=3.0.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
thefuzz>=0.20.0
python-Levenshtein>=0.21.0
jinja2>=3.1.0
